semantic_error: let it own its chain
This patch simplifies the way semantic_error chains are used by allowing
the parent object to own its chain. Upon setting the chain, the parent
creates a copy and keeps it secret, to be de-allocated upon destruction.

The patch also modifies the semantic_error constructor to allow the
chain to be also set at the same time.