Remember which rpms have been checked in the session
When the kernel or any userspace file is missing debuginfo, we run an
rpm query so packages can be suggested.  If the user tries many such
probes on the same target, it's a waste to repeat the same query.  Now
we remember which targets have already been checked in the session.

This was seen on a simple 'stap -l syscall.*', which took much longer to
run when debuginfo was missing than when present.  With this patch, the
first syscall miss will lead to an rpm query, but each following miss
will know it's already been done.