Add a new command line option to enable print backtraces under some conditions.
Most of the times, when we use bkl.stp to investigate BKL issues, we've found useful to have
the backtrace of the function that holds the BKL. This patch adds a new command line option,
"backtrace=N", that when used enables the use of print_backtrace(). The value of <N>, if present,
is taken as useconds, and it is the limit of time the task or tasks have been waiting to acquire
the lock i.e.:

  # stap blk.stp backtrace=100

That would enable printing of backtrace on all tasks holding the kernel
big lock for more than 100us.

Signed-off-by: Jose Castillo <jcastillo@redhat.com>