callee probes: truly improve derivation for better listing
The previous attempt to do this (commit 80a63be) was hacky and was thus
reverted (see previous commit). In this commit, we try to do it
properly.

--- original commit msg ---

Previously, doing stap -L for function("foo").callee("*") would result
in listings directly for function("bar1"), function("bar2"), etc...

With this patch, we add an intermediate step in the derivation chain:
function("foo").callee("*") --> function("foo").callee("bar1")

This will cause -L to display this derivation instead, which is closer
to what users would expect for a function("foo").* probe.

To achieve this, we do three things:

1. We make func_info and inline_instance_info subclasses of a single
   base_func_info class for uniform treatment when called from
   query_cu() as well as cleaner passing of name/file/line/addr between
   iterate_over_callees() and query_callee(). This can also help in
   other cases (to be assessed...).
2. We remember and pass down the caller information in
   iterate_over_callees(), so that query_callee() has it on hand when
   creating the .function component.
3. In query_callee(), we create an intermediate probe point and a new
   dwarf_query based on that probe point to create the final derived
   probe.