BZ1073605 addendum, don't blacklist empty matches
For example, plt and symtab function probes have an empty filename, but
they should not be blocked for that.  This patch adds an impossible ".^"
pattern as a placeholder in those blacklist patterns that aren't used
yet.  Also add semok/semko tests to be sure probing "_start" is allowed
on entry but not on return.