PR 16645: Unhelpful feedback when stap fails
Under some conditions, stap automatically sets options which
have not been explicitely set, in an attempt to succeed where
it is known that translation would fail without these settings.
Examples are the automatic setting of the --privilege level and
the automatic request for translation via a compile-server for
unprivileged users.

In these cases, when translation still fails it would be helpful
to the user to explain what settings were changed automatically
and why.

It is also helpful to point the user at the appropriate documention
when futher configuration/setup  on his part is required.

This change does the above for --privilege and --use-server.