RHBZ 1075685 - tap-server trusting issues
If a new server to be trusted was selected by address + port,
and there was no avahi assistance available, or the server was not known
to avahi, its certificate was not added to the database of
trusted servers. This was because the client mishandled the
case of an empty certificate serial number and acted as if
the new server was already trusted.

Fix this by not searching the NSS database of trusted certificates
for a match against an empty certificate serial number.