PR15999: strengthen null die checking
In commit b78a0fb, we allow $$vars/$$param/$$locals/$$return to expand
to an empty string if the scope_die is missing, such as is the case for
symtab-derived functions.

This patch ensures that we do not try to resolve normal context
variables, such as $var or $return, which absolutely require a
scope_die, and for which we previously SIGABRT'ed.

We also strengthen getscopes(), which only pointer-checked the
scope_die, rather than using null_die() to value-check it.