RHBZ1065757: don't delete objects that are in use
Catching a semantic_error does not mean that nothing was derived. This
can occur in the case of probe aliases, where the first alias_name was
derived, but the second wasn't. In those cases, the first derived probe
uses those created objects, so corruption may result if we delete them.

We instead rely on the size of the results vector to determine if it is
safe to delete the objects.