Read the perf process from -c during pass2
When a bare .process is used, it's value is inferred from -c CMD.  This
must be checked during pass2, or else a changed -c CMD will not trigger
a hash change, and an incorrect cached module will be used.

This patch also unifies the wordexp argv[0] parsing of -c CMD into a
shared systemtap_session::cmd_file().