Experiment with forking the compilation passes
This moves the passes_0_4() compilation into a forked subprocess.  This
guarantees that any resources used (heap, files, etc.) are not held
throughout pass-5 while we actually run the script.  Even just for the
heap, this is the difference of 10s or 100s of MB RSS while we run,
depending on the complexity of the script, to just around 3MB RSS.