PR16844: Initial adaptation for kernel 3.15 tracepoints
This hides the tracepoint version adaptations in stp_tracepoint.h.  For
older kernel versions, that header is sufficient.  With 3.15 we have to
do extra bookkeeping ourselves, so it pulls in stp_tracepoints.c, which
is borrowed almost verbatim from lttng.  It seems to work well, but we
can still decide if we want to take a different approach.