statement.exp: adapt for RHEL5
In GCC < 4.4, the decl_line is the line holding the opening brace, not the
line with the identifier. Since the decl_line is the basis for relative
line number statement probes, this discrepancy was causing statement.exp
to have some RELATIVE subtest failures.

Add another empty line so that main@statement.c+4 always falls on an
empty line, regardless of which is the decl_line.