PR16406: fix build-id false mismatch for userspace-program vs. kernel-module
The translate.cxx-emitted _stp_module array has two name fields, .name
and .path, which were ambiguous as to whether each was a kernel module
or a userspace executable.  The .name string was just the basename,
and it was used in the near-startup-time _stp_module_check().  If the
userspace program matched a kernel module, bangorama (a false build-id
mismatch error at startup) .. unless --all-modules was given, in
which case the kernel modules won the duplication race.

The correction has a couple of parts.  First, the _stp_module .name is
made the same as the fully-/-qualified .path for userspace programs.
Second, _stp_module_check checks all kernel+module bits it knows
about, not just the first one.  Hand-tested with a variety of runs,
and with the *myproc.exp test, which is now forced to try to conflict.

* translate.cxx (dump_unwindsym_ctx): Make userspace _stp_module->name's
  fully-/-qualified.
* runtime/sym.c (_stp_module_check): Check all and only kernel-space
  build-ids.  Filter by _stp_module->name.
  (stp_kmodule_check, stp_usermodule_check): WARN_ON wrong-polarity
  incoming module/path name.
* testsuite/.../unprivileged_myproc.exp: Run 'modprobe loop'