pyexample.stp: split into py2example.stp and py3example.stp
In order to be compatible on systems with only python2 installed (or
only python3 installed), we split the pyexample.stp script into two
versions.

We also add a test_support line in the meta files so that check.exp
marks them as untested if either/both python versions are not present.

In the tapsets, we make py2/3 aliases that the scripts then use, but
also create general ones that refer to both the 2 and 3 variants.

The python3 version requires -g --suppress-time-limits, otherwise it's
very likely that a 'overhead exceeded' ERROR.

- py2example.stp: python 2 version of the script
- py2example.meta: meta file for python 2 version, with test_support
  check for python 2 specific SDT markers
- py3example.stp: python 3 version of the script
- py3example.meta: meta file for python 3 version, with test_support
  check for python 3 specific SDT markers
- tapset/python2.stp: make python2 specific aliases and alias the
  general ones to that
- tapset/python3.stp: make python3 specific aliases and alias the
  general ones to that