library probes: strengthen library path resolution
One of two things happens to library probes the first time we enter
dwarf_builder::build():
   - If the full path can be resolved using find_executable(), then we
     would right away switch focus to the library, rather than the
     process.
   - Otherwise, we would focus on the process and then use
     iterate_over_libraries() to find a library match.

These decisions are made in two spots: in dwarf_builder::build(), prior
to creating the dwflpp object, and in base_query::base_query().

When we resort to iterate_over_libraries(), new probes are created for
each matching library. And since the path inserted in those new probes
is the full path, find_executable() will find it upon recursing into
dwarf_builder::build().

One issue is that if find_executable() succeeds on the first time, we
will focus on the library, regardless of whether the library is NEEDED
or not. Whereas if find_executable() fails, only NEEDED libraries could
ever be picked up by iterate_over_libraries().

This is mostly fine, except that it creates an inconsistency. For
example, let's say executable 'bar' depends on 'libfoo.so' located in
/lib/. Doing a "libfoo.so" library probe would prioritize a libfoo.so
file in the current directory before the one in /lib/, which is the
'real' one. On the other hand, doing "libfoo.*" would match the
libfoo.so file in /lib/ because of the glob char.

It also fixes the potential issue of having a fully specified library
path with a bad process path: we would focus on the library yet in
query_module() try to iterate_over_libraries() on the library, thinking
that the library was not resolved.

In this series of patches, we change the behaviour as follow
   - Only switch to the library on the first entry of
     dwarf_builder::build() if the path is already fully resolved.
     Otherwise, always use iterate_over_libraries().
   - If iterate_over_libraries() fails, then try a direct
     find_executable() invocation, and warn that the found library is
     not a NEEDED one.

This patch does the first part only. The second part is done in the next
patch.