callee.exp: simplify shlib testing by using -rpath
We previously started stap separately from the target process because we
needed to set the LD_LIBRARY_PATH variable for it to run properly.

However, if we use the linker option -rpath to add the current directory
to the library path (as is done also in unprivileged_myproc.exp), this
is no longer necessary, thus greatly simplifying the testcase.