PR17027: adapt pr16806.exp for RHEL6
This test failed on RHEL6 for two reasons:
- The 'timeout' program doesn't support non-integer arguments
- We assumed the uprobes module was already inserted.

Fix the first by changing the timeout to 1s (I verified that this
doesn't change the lethality of the test under the right conditions).

Fix the second by doing a sample run first to allow stap the chance to
self-compile and pass the uprobes module to staprun -u.