Let @cast reuse modules even without a cache
It's sometimes necessary for autocast that separate @cast expressions
resolve to the same type, but that requires them to use the same underlying
module.  When the cache was disabled, we were compiling a new module for
every @cast encountered.

Now the @cast visitor remembers what modules it has used for a given header
in this session, which will both help performance and make autocast type
comparisons work directly.