initscript: check script names
Because script names are the basis for some variable names (e.g.
script_name_REQ), they need to adhere to bash variable naming rules.
These are almost the same rules that we enforce for stap -m, so we just
fail 'earlier'. See also RHBZ1109084.

This patch ensures that scripts with invalid names are skipped and
logged.