PR17052: Use RCU-sched for freeing contexts
A context is grabbed from the _stp_ctl_work_callback timer, but it
doesn't actually need it for anything other than preventing a lock
recursion.  Since this timer is still running while we are freeing the
contexts, there's a chance to race on the existence of contexts.

Using rcu_assign_pointer, rcu_dereference_sched, and synchronize_sched,
we can make sure that anyone who might have seen the existing contexts
has finished before we free them.