listing_mode.exp: use stap instead of nm to get addrs
We previously relied on nm to retrieve the addresses of functions in
the executable and library. This causes issues on ppc64, in which nm
simply returns the function descriptor rather than the actual text
address.

We now instead rely on stap itself to retrieve function addresses. This
should fix all of the FAILs of this testcase on the ppc64 platform.