modules_out_of_tree.exp: adapt for RHEL5
We put the opening braces on the same line as the function declaration
so that the reported lineno in listing mode is the same on RHEL5 GCC
(which considered the opening brace as the decl_line).

We also skip the probing part by checking for module refresh support by
using the new module_refresh_p proc.