on-the-fly: use a background timer to schedule work
Calling schedule_work() is not always safe from some contexts (e.g. when
tracing/probing the internals of workqueues themselves).

We remove the code which previously called schedule_work() in the common
epilogue of all probe types. We will need to vet on a case-by-case basis
which probe types are safe.

Meanwhile, we implement a background timer which simply checks if
schedule_work() needs to be called.