initscript: use --save-uprobes instead of -u
Rather than masking the -u option of stap (unoptimized mode), use the
actual option which we use with stap (--save-uprobes). This will allow
us to remain backwards-compatible if we later on automate the uprobes
saving by always turning on --save-uprobes.

So now, users have to do the following if they require userspace
probing:

  # cat /etc/systemtap/conf.d/example.conf
  example_OPT=--save-uprobes