initscript: automatically detect uprobes dependency
systemtap-initscript scripts that rely on uprobes must be configured
with the --save-uprobes option.  This option saves the generated
uprobes.ko module and loads it when running the script.

The uprobes dependency information is actually available at compile time
so we can autodetect as follows:
1. Check if uprobes.ko was generated during compile
2. When uprobes.ko was generated, touch <name>.uprobes in the cache
   directory.
3. Add the staprun -u option if <name>.uprobes exists

It is no longer necessary to specify the --save-uprobes option in the
initscript configuration file, although doing so allowed.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>